user-service/requirements.txt
alembic==1.10.2
autoenv==1.0.0
blinker==1.5
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
click==8.1.3
cryptography==40.0.1
dnspython==2.3.0
dominate==2.7.0
Flask==2.2.3
Flask-DotEnv==0.1.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-Uploads==0.2.1
Flask-WTF==1.0.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
marshmallow==3.19.0
passlib==1.7.4
protobuf==4.22.1
pycparser==2.21
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.7
urllib3==1.26.15
visitor==0.1.3
Werkzeug==2.2.3

